+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## **Unreleased**
 **Hearthstone**
-- Fixed The Kingslayer not Highlighting Incindius' Eruptions.
 - Added cards offered by Waveshaping to the bottom of the deck.
 - Added cards discovered by Kaldorei Cultivator to the bottom of the deck.
+- Added Timethief Rafaam Counter.
 - Fixed crashes when Murozond, Unbounded's attack was set to infinity. HDT is now capable of handling all its power.
+- Fixed The Kingslayer not Highlighting Incindius' Eruptions.
 - Fixed fabled bundled cards not appearing as Legendary cards in the decklist.
 - Fixed King Llane not appearing in opponent's decklist.